requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cffi==1.15.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
libnacl==1.8.0
multidict==6.0.2
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.20.0
typing_extensions==4.2.0
yarl==1.7.2
